+++ libcloud/compute/types.py
@@ -184,6 +184,10 @@ class NodeState(object):
     :cvar TERMINATED: Node is terminated. This node can't be started later on.
     :cvar STOPPED: Node is stopped. This node can be started later on.
     :cvar PENDING: Node is pending.
+    :cvar STOPPED: Node is stopped.
+    :cvar SUSPENDED: Node is suspended.
+    :cvar ERROR: Node is an error state. Usually no operations can be performed
+                 on the node once it ends up in the error state.
     :cvar UNKNOWN: Node state is unknown.
     """
     RUNNING = 0
@@ -192,6 +196,8 @@ class NodeState(object):
     PENDING = 3
     UNKNOWN = 4
     STOPPED = 5
+    SUSPENDED = 6
+    ERROR = 7
 
 
 class Architecture(object):